linux-raspberrypi: Add a 4.1 linux kernel with vc4 support
This adds Eric Anholt's WIP kernel with dri/kms/3d support for the GPU on
the rpi2. Adding a recipe that tracks this kernel, will make it easier to
test this setup.

This recipe should only used for testing purposes for now, so it will not
be set as the default but can be chosen by changing the default providers.

Signed-off-by: Derek Foreman <derekf@osg.samsung.com>
Signed-off-by: Javier Martinez Canillas <javier@osg.samsung.com>