rpi-default-providers: Document how to switch providers for vc4 gfx stack
The Raspberry Pi boards can use one of two graphics stacks: The "userland"
user-space driver or the VC4 drm/kms kernel driver. This patch documents
how the user user can choose if the VC4 driver has to be used instead and
set the right preferred providers for that case.

The README a section that explains the RPi can different graphics stacks
and that the user can choose by manually changing providers.

Signed-off-by: Derek Foreman <derekf@osg.samsung.com>
Signed-off-by: Javier Martinez Canillas <javier@osg.samsung.com>